#0: moving ViT from experimental to demos
