+++ models/demos/wormhole/vit/demo/demo_vit_ttnn_inference_perf_e2e_2cq_trace.py
@@ -9,7 +9,7 @@ from loguru import logger
 
 import ttnn
 
-from models.experimental.functional_vit.tt import ttnn_optimized_sharded_vit_wh
+from models.demos.vit.tt import ttnn_optimized_sharded_vit_wh
 from models.utility_functions import is_blackhole
 
 from models.utility_functions import (
