+++ tests/ttnn/integration_tests/vit/test_accuracy_ttnn_optim_interleaved_vit.py
@@ -15,8 +15,8 @@ from transformers import AutoImageProcessor
 
 import ttnn
 
-from models.experimental.functional_vit.tt import ttnn_optimized_interleaved_vit
-from models.experimental.vit.vit_helper_funcs import get_data_loader, get_batch
+from models.demos.vit.tt import ttnn_optimized_interleaved_vit
+from models.demos.vit.vit_helper_funcs import get_data_loader, get_batch
 
 from ttnn.model_preprocessing import preprocess_model_parameters
