+++ tests/ttnn/integration_tests/vit/test_torch_functional_vit_highres.py
@@ -10,7 +10,7 @@ import transformers
 
 from ttnn.model_preprocessing import preprocess_model_parameters
 
-from models.experimental.functional_vit.reference import torch_functional_vit
+from models.demos.vit.reference import torch_functional_vit
 from models.utility_functions import torch_random, is_blackhole, is_wormhole_b0
 
 from tests.ttnn.utils_for_testing import assert_with_pcc
