+++ tests/ttnn/integration_tests/vit/test_ttnn_optimized_sharded_vit.py
@@ -13,8 +13,8 @@ from transformers import AutoImageProcessor
 import ttnn
 from ttnn.model_preprocessing import preprocess_model_parameters
 
-from models.experimental.functional_vit.tt import ttnn_optimized_sharded_vit
-from models.experimental.functional_vit.reference import torch_functional_vit
+from models.demos.vit.tt import ttnn_optimized_sharded_vit
+from models.demos.vit.reference import torch_functional_vit
 from models.utility_functions import torch_random, is_wormhole_b0, is_blackhole
 
 from tests.ttnn.utils_for_testing import assert_with_pcc
